backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
pydantic==2.9.2
pydantic-extra-types==2.10.1
pydantic-core==2.23.4
email-validator==2.2.0
pydantic-settings==2.6.1
python-multipart==0.0.12
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
bcrypt==4.2.0
alembic==1.13.3

